4/4/24 RX, Z, Names + more + sql error fix 4/6/24
work on holocrypts and Rowsdowers started.
I also reodering stats, so it will be easier to find them
I spaced out the numbers to make then in line with each other, this give me space to input numbers later. 
Next update will see more HP estimated, and more work on Earth room. from now on i will not call it earth room, but it will be known as spider room, after the 100+ spiders that need to place.
And maybe holocripts will be in, unknown for Rowsdowers, most likely a temp spot till i find the real spot.