Event Rework
This rework move us off the old format of a sql file for every zone (Winterfest/Shade's Eve, are the only two that apply to all zones [maybe Arcterra is missing them]) for every event, plus add that some people don't want the Halon Ring update for Illium & Thayd, so we get much more files. In the old format you need to run the event sql for that zone, then run the non event sql to remove it, the issue with this is that you would have 53 sql files (maybe more) for events. This rework solves this, you don't need to run the zone's sql and only thing you need to do is run the event's sql, and to remove it you run the  Remove any event.sql this will only remove the event. This is still a WIP, I didn't finish placing everything, and remember if the server does not tell the client the event is live, the event won't work, the objects and NPCs will be here, but they won't give quests and the lasers won't turn on, etc, try Krimmin's Shade's Eve build, and edit it to have some other events running. If you have any sniffs of events running in Thayd, Illium, zPrix in Whitevale, Celestion, or zones while Shade's Eve or Winterfest is running, send me a DM on discord   michaeln23  I hope this rework will make server admin's life easier.